Modularize using MVC pattern
