+++ Modules/Python/Flask and MySQL/Core/Users CRUD/flask_app/__init__.py
@@ -1,0 +1,4 @@
+from flask import Flask
+
+app = Flask(__name__)
+app.secret_key = "blaBlAlL/a1234 aA./HGH"
