+++ Modules/Python/Flask and MySQL/Core/Users CRUD/flask_app/models/user.py
@@ -1,5 +1,5 @@
 # import the function that will return an instance of a connection
-from mysqlconnection import connectToMySQL
+from flask_app.config.mysqlconnection import connectToMySQL
 # model the class after the user table from our database
 class User:
     DB = "users_crud"
