+++ Modules/Python/Flask and MySQL/Core/Users CRUD/flask_app/templates/user.html
